requirements.txt
tcrutils>=12.0.0,<13.0.0
hikari>=2.0.0
hikari-miru>=4.0.0,<5.0.0
hikari-arc>=1.2.1
hikari-lightbulb>=2.3.5
attrs>=23.1.0
